==> RAnalysis/Scripts/Transcriptomics/07_KEGG.Rmd <==
---
title: "KEGG Enrichment WGCNA"
author: "Samuel Gurr"
date: "2024-01-24"
output: html_document
---

## Load libraries, set up
```{r setup, include=FALSE}
# LOAD PACKAGES
library(clusterProfiler)
library(KEGGREST)
library(tidyr)
library(stringr)
library(forcats)
library(ggplot2)
library(scales)
library(ape)
library(data.table)
library(tidyverse)
# library(fBasics)
library(dplyr)
# library(BaseSet)

# set working directory
knitr::opts_knit$set(root.dir = "C:/Users/samjg/Documents/Github_repositories/Airradians_CellularMolecular_OA/RAnalysis/Output/Transcriptomics") # sets the working
```


## Load data

What data are we targetting?

(1) Cohort 

* DESEq2 results for within pCO2h. Results are the overlapped genes between three models low v. moderate, low v. high and moderate v. high

- low expression low pCO2h 

- low expression moderate pCO2h

- low expression high pCO2h

(2) Challenge

* WGCNA results for pCO2h x pCOex. We have four modules with significant correlations with treamtent, pink, magenta, red, and brown. Pink and magenta (high expression high pCO2h) were used to investigate frontloaded genes. Red and brown (low expression high pCO2h) were investigated for 'costly activation'

- full module analysis 
  
  - red
  - brown
  - pink 
  - magenta

- subset (these are the most intersting, likley main manuscrupt material here!)

  - frontloaded (using pink and magenta)
  
  - costly activation (using red and brown)


```{r  load data}


# (1) Cohort data - load and mutate new KEGG id column 
# ( note KEGG ids are simply the geneID with cvn: before it, remove LOC)

# WGCNA cohort data
Cohort_ModuleMembership      <- read.csv(file="WGCNA/cohorts/Cohorts_SignificantModules_Annotated.csv",
                                        sep=',', header=TRUE) # moduleColor %in% c("tan", "grey60", "midnightblue", "purple", "brown", "black"))

# DESEQ2 cohort data
CohortRes_LowExp.LowpCO2h <- read.csv(file="DESeq2/Venn_intersections/Low_expression_LowpCO2.csv",
                                        sep=',', header=TRUE) %>% 
                            dplyr::mutate(Cvirginica_KEGGID = paste0('cvn:',gsub(".*LOC", "", blastxEval_CvirgProteinID))) %>% 
                            dplyr::filter(!Cvirginica_KEGGID %in% 'cvn:NA')

CohortRes_LowExp.ModpCO2h <- read.csv(file="DESeq2/Venn_intersections/Low_expression_ModeratepCO2.csv",
                                        sep=',', header=TRUE) %>% 
                            dplyr::mutate(Cvirginica_KEGGID = paste0('cvn:',gsub(".*LOC", "", blastxEval_CvirgProteinID))) %>% 
                            dplyr::filter(!Cvirginica_KEGGID %in% 'cvn:NA')

CohortRes_LowExp.HighpCO2h <- read.csv(file="DESeq2/Venn_intersections/Low_expression_HighpCO2.csv",
                                        sep=',', header=TRUE) %>% 
                            dplyr::mutate(Cvirginica_KEGGID = paste0('cvn:',gsub(".*LOC", "", blastxEval_CvirgProteinID))) %>% 
                            dplyr::filter(!Cvirginica_KEGGID %in% 'cvn:NA')


# (2) Challenge data
# ( note KEGG ids are simply the geneID with cvn: before it, remove LOC)
ChallengeRes_FullModules      <- read.csv(file="WGCNA/challenge/Challenge_SignificantModules_Annotated.csv",
                                        sep=',', header=TRUE) 
  

ChallengeRes_Frontloaded      <- read.csv(file="WGCNA/challenge/Frontloaded_Activated/UPmodulePinkMagenta_Frontloaded.csv",
                                   sep=',', header=TRUE) %>%  dplyr::rename('Airradians_TranscriptID' =
                                                                                   'Airradians.TranscriptID')%>% 
                            dplyr::mutate(Cvirginica_KEGGID = paste0('cvn:',gsub(".*LOC", "", Cvirginica_GeneID))) %>% 
                            dplyr::filter(!Cvirginica_KEGGID %in% 'cvn:NA')
# originally 711 and now 612  genes with NAs removed (Airradians genes without Cvirg hit for KEGG annotation)
#((711-612)/711)*100 # 13.92405% wihtout KEGG

ChallengeRes_Activated      <- read.csv(file="WGCNA/challenge/Frontloaded_Activated/UPmodulePinkMagenta_Activated.csv",
                                   sep=',', header=TRUE) %>%  dplyr::rename('Airradians_TranscriptID' =
                                                                                   'Airradians.TranscriptID')%>% 
                            dplyr::mutate(Cvirginica_KEGGID = paste0('cvn:',gsub(".*LOC", "", Cvirginica_GeneID))) %>% 
                            dplyr::filter(!Cvirginica_KEGGID %in% 'cvn:NA')


ChallengeRes_CostlyActivation <- read.csv(file="WGCNA/challenge/Frontloaded_Activated/DOWNmoduleBrownRed_CostlyActivation.csv",
                                   sep=',', header=TRUE) %>%  dplyr::rename('Airradians_TranscriptID' =
                                                                                   'Airradians.TranscriptID')%>% 
                            dplyr::mutate(Cvirginica_KEGGID = paste0('cvn:',gsub(".*LOC", "", Cvirginica_GeneID))) %>% 
                            dplyr::filter(!Cvirginica_KEGGID %in% 'cvn:NA')
# originally 866 and now 687  genes with NAs removed (Airradians genes without Cvirg hit for KEGG annotation)
#((866-687)/866)*100 # 20.66975% wihtout KEGG
```

# KEGG enrichment
- enrichKEGG using clusterProfiler

## (1.1) Cohort (WGCNA, significnat modules p < 0.05)

* note! no enrichment for modules tan, grey60 and midnightblue
```{r Cohort for loop all sig modules - run KEGG}

# call the module colors 
cohort_modcolor <- as.data.frame(unique(Cohort_ModuleMembership$moduleColor) %>% na.omit())
names(cohort_modcolor)[1] <- "color"

# run for loop - enrichKEGG with output csv, unlisted genes master csv. rich factor plots
for(i in 1:nrow(cohort_modcolor)) {
  
  loop_df         <- Cohort_ModuleMembership %>% filter(moduleColor %in% cohort_modcolor[i,])
  entrezID_vector <- as.vector(as.numeric(gsub('.*cvn:', '',loop_df$Cvirginica_KEGGID)))
  KEGG_cvirg      <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05)
  
  if(nrow(as.data.frame(head(KEGG_cvirg))) == 0 ) { # if no KEGG enrichment 
    # do nothing
  } else { # in all other cases, meaning there is at least 1 enriched pathway
    df              <- as.data.frame(head(KEGG_cvirg))
    rownames(df)    <- c()
    KEGGoutput      <- as.data.frame(do.call(cbind.data.frame, df)) %>% 
                                                  dplyr::mutate(
                                                                Rich_factor  = (  (as.numeric(sub("/.*", "", GeneRatio))) / 
                                                                            (as.numeric(sub("/.*", "", BgRatio)))),
                                                                Type = paste0('Cohort WGNCA module ',modcolor[i,])) %>% 
                                                  dplyr::rename(KEGGID_pathway = ID,
                                                                pathway.name   = Description,
                                                                Gene.IDs       = geneID,
                                                                Num.genes.exp  = Count) %>% 
                                                  dplyr::mutate(Num.genes.all  = (as.numeric(sub("/.*", "", BgRatio))), 
                                                                log10_pvalue   = abs(log10(pvalue)),
                                                                pathway.name = sapply(strsplit(pathway.name, " - Crassostrea"),
                                                                                      "[",1)) %>% # ommit the unneeded pathway string
                                                  dplyr::select('Type', 'KEGGID_pathway', 'pathway.name',
                                                                'Num.genes.all', 'Num.genes.exp', 'Gene.IDs',
                                                                'Rich_factor', 'pvalue', 'log10_pvalue', 'qvalue') %>% 
                                                  arrange(desc(as.numeric(Rich_factor)))
    write.csv(KEGGoutput, file = paste0("WGCNA/cohorts/clusterProfiler/Module_",modcolor[i,],"_KEGG.csv", sep ='')) 
    
    # Rich Factor plot 
    plot<- KEGGoutput %>%  
              ggplot(aes(x=reorder(pathway.name, Rich_factor), y= Rich_factor)) + 
              geom_point( aes(col=qvalue, size=Num.genes.exp)) +   # Draw points
              geom_segment(aes(x=pathway.name, 
                               xend=pathway.name, 
                               y=min(Rich_factor), 
                               yend=max(Rich_factor)),  
                               linetype=NA, 
                               size=0) +   # Draw dashed lines
              labs(title=paste0('Cohort WGNCA module ',modcolor[i,]), 
                   x = "Pathway",
                   y = "Rich Factor") +
              theme_bw() +
              coord_flip()
    pdf(paste("WGCNA/cohorts/clusterProfiler/Module_",modcolor[i,],"_RichFactorPlot.pdf", sep =''),width=8, height=6)
    print(plot)
    dev.off()
                     
    # stringsplit and unnest for a data set of genes and IDs associated with each pathway 
    KEGGoutput$Gene.IDs  <- as.vector(strsplit(as.character(KEGGoutput$Gene.IDs), "/"))
    KEGGoutput_unnest                     <- unnest(KEGGoutput, Gene.IDs)
    KEGGoutput_unnest$Cvirginica_KEGGID   <- paste("cvn:", KEGGoutput_unnest$Gene.IDs, sep='')
    KEGGoutput_allgenes                   <- merge(KEGGoutput_unnest, 
                                                    (Cohort_ModuleMembership %>% 
                                                       select(Protein_name, Cvirginica_KEGGID, 
                                                              Cvirginica_TranscriptID, Airradians_TranscriptID)),
                                                    by='Cvirginica_KEGGID') %>% 
                                                                           group_by(pathway.name) %>% 
                                                                           arrange(Protein_name, .by_group = TRUE) %>%
                                                                           unique()
    
    write.csv(KEGGoutput_allgenes, file = paste("WGCNA/cohorts/clusterProfiler/Module_",modcolor[i,],"_KEGG_unlisted.csv", sep ='')) 

  } # end of if else statement
  
  
} # close for loop



```

* run tan + grey60 and brown + black since these had the same eigengene pattern with pCO2 history
  - result: still no enrichment for tan + grey 60! enrichment for brown + black
```{r Cohort overlap modules of same pattern - run KEGG}

# Now run KEGG for modules combined by tretment correlation
entrezID_vector_grey_tan <- as.vector(as.numeric(gsub('.*cvn:', '',
                                             (Cohort_ModuleMembership %>%  
                                               filter(moduleColor %in% c('tan', 'grey60')))$Cvirginica_KEGGID
                                             )))
KEGG_cvirg_grey_tan      <- enrichKEGG(gene = entrezID_vector_grey_tan, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05)
# NO ENRICHMENT FOR TAN and GREY60!!

entrezID_vector_black_brown <- as.vector(as.numeric(gsub('.*cvn:', '',
                                             (Cohort_ModuleMembership %>%  
                                               filter(moduleColor %in% c('black', 'brown')))$Cvirginica_KEGGID
                                             )))
KEGG_cvirg_black_brown      <- enrichKEGG(gene = entrezID_vector_black_brown, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05)
KEGGoutput_black_brown      <- as.data.frame(KEGG_cvirg_black_brown) %>% 
                                                  dplyr::mutate(
                                                                Rich_factor  = (  (as.numeric(sub("/.*", "", GeneRatio))) / 
                                                                            (as.numeric(sub("/.*", "", BgRatio)))),
                                                                Type = 'Cohort WGNA modules brown + black') %>% 
                                                  dplyr::rename(KEGGID_pathway = ID,
                                                                pathway.name   = Description,
                                                                Gene.IDs       = geneID,
                                                                Num.genes.exp  = Count) %>% 
                                                  dplyr::mutate(Num.genes.all  = (as.numeric(sub("/.*", "", BgRatio))), 
                                                                log10_pvalue   = abs(log10(pvalue)),
                                                                pathway.name = sapply(strsplit(pathway.name, " - Crassostrea"),
                                                                                      "[",1)) %>% # ommit the unneeded pathway string
                                                  dplyr::select('Type', 'KEGGID_pathway', 'pathway.name',
                                                                'Num.genes.all', 'Num.genes.exp', 'Gene.IDs',
                                                                'Rich_factor', 'pvalue', 'log10_pvalue', 'qvalue') %>% 
                                                  arrange(desc(as.numeric(Rich_factor)))
write.csv(KEGGoutput_black_brown, file = paste0("WGCNA/cohorts/clusterProfiler/Modules_brown_black_KEGG.csv", sep =''))
# Plot
plot<- KEGGoutput_black_brown %>%  
          ggplot(aes(x=reorder(pathway.name, Rich_factor), y= Rich_factor)) + 
          geom_point( aes(col=qvalue, size=Num.genes.exp)) +   # Draw points
          geom_segment(aes(x=pathway.name, 
                           xend=pathway.name, 
                           y=min(Rich_factor), 
                           yend=max(Rich_factor)),  
                           linetype=NA, 
                           size=0) +   # Draw dashed lines
          labs(title="Cohort WGCNA brown + black", 
               x = "Pathway",
               y = "Rich Factor") +
          theme_bw() +
          coord_flip()
pdf(paste("WGCNA/cohorts/clusterProfiler/Module_brown_black_RichFactorPlot.pdf", sep =''),width=8, height=6)
print(plot)
dev.off()
                     
# stringsplit and unnest for a data set of genes and IDs associated with each pathway 
KEGGoutput_black_brown$Gene.IDs  <- as.vector(strsplit(as.character(KEGGoutput_black_brown$Gene.IDs), "/"))
KEGGoutput_black_brown_unnest                     <- unnest(KEGGoutput_black_brown, Gene.IDs)
KEGGoutput_black_brown_unnest$Cvirginica_KEGGID   <- paste("cvn:", KEGGoutput_black_brown_unnest$Gene.IDs, sep='')
KEGGoutput_black_brown_allgenes                   <- merge(KEGGoutput_black_brown_unnest, 
                                                (Cohort_ModuleMembership %>% 
                                                   select(Protein_name, Cvirginica_KEGGID, 
                                                          Cvirginica_TranscriptID,Airradians_TranscriptID)),
                                                by='Cvirginica_KEGGID') %>% 
                                                                       group_by(pathway.name) %>% 
                                                                       arrange(Protein_name, .by_group = TRUE) %>%
                                                                       unique()

write.csv(KEGGoutput_black_brown_allgenes, file = paste("WGCNA/cohorts/clusterProfiler/Module_brown_black_KEGG_unlisted.csv", sep ='')) 
```

## (1.2) Cohort (DESeq2, overlapped genes)

* there are three cases at which the vast majorty of DEGs are overlapped between binary models,
below are the each, and we investigate KEGG pathway enrichment

### Low Expression by LowpCO2h (low v. moderate and low v. high, overlapped downreg DEGs in both models)

**input:** CohortRes_LowExp.LowpCO2h

```{r Cohort Low Expression by LowpCO2h  DESEq2 results - run KEGG}

entrezID_vector <- as.vector(as.numeric(gsub('.*cvn:', '',CohortRes_LowExp.LowpCO2h$Cvirginica_KEGGID)))
KEGG_cvirg      <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05) 
# ZERO ENRICHED TERMS FOUND
?enrichKEGG
```

### Low Expression by ModeratepCO2h (low v. moderate and moderate v. high, overlapped DEGs for a sureg pattern by moderate in both models)

**input:** CohortRes_LowExp.ModpCO2h

```{r Cohort Low Expression by ModeratepCO2h  DESEq2 results - run KEGG}

entrezID_vector <- as.vector(as.numeric(gsub('.*cvn:', '',CohortRes_LowExp.ModpCO2h$Cvirginica_KEGGID)))
KEGG_cvirg     <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05) 
df              <- as.data.frame(head(KEGG_cvirg))
rownames(df)    <- c()
KEGGoutput      <- as.data.frame(do.call(cbind.data.frame, df)) %>% 
                                                dplyr::mutate(
                                                              Rich_factor  = (  (as.numeric(sub("/.*", "", GeneRatio))) / 
                                                                          (as.numeric(sub("/.*", "", BgRatio)))),
                                                              Type = 'Cohort low expression moderate pCO2') %>% 
                                                dplyr::rename(KEGGID_pathway = ID,
                                                              pathway.name   = Description,
                                                              Gene.IDs       = geneID,
                                                              Num.genes.exp  = Count) %>% 
                                                dplyr::mutate(Num.genes.all  = (as.numeric(sub("/.*", "", BgRatio))), 
                                                              log10_pvalue   = abs(log10(pvalue)),
                                                              pathway.name = sapply(strsplit(pathway.name, " - Crassostrea"),
                                                                                    "[",1)) %>% # ommit the unneeded pathway string
                                                dplyr::select('Type', 'KEGGID_pathway', 'pathway.name',
                                                              'Num.genes.all', 'Num.genes.exp', 'Gene.IDs',
                                                              'Rich_factor', 'pvalue', 'log10_pvalue', 'qvalue') %>% 
                                                arrange(desc(as.numeric(Rich_factor))) 
                    
write.csv(KEGGoutput, file = paste("DESeq2/Venn_intersections/clusterProfiler/Low_expression_ModeratepCO2_KEGG.csv", sep ='')) 
                    
# Plot
plot<- KEGGoutput %>%  
          ggplot(aes(x=reorder(pathway.name, Rich_factor), y= Rich_factor)) + 
          geom_point( aes(col=qvalue, size=Num.genes.exp)) +   # Draw points
          geom_segment(aes(x=pathway.name, 
                           xend=pathway.name, 
                           y=min(Rich_factor), 
                           yend=max(Rich_factor)),  
                           linetype=NA, 
                           size=0) +   # Draw dashed lines
          labs(title="Low Expression Moderate pCO2h", 
               x = "Pathway",
               y = "Rich Factor") +
          theme_bw() +
          coord_flip()
pdf(paste("DESeq2/Venn_intersections/clusterProfiler/Low_expression_ModeratepCO2_RichFactorPlot.pdf", sep =''),width=8, height=6)
print(plot)
dev.off()
                     
# stringsplit and unnest for a data set of genes and IDs associated with each pathway 
KEGGoutput$Gene.IDs  <- as.vector(strsplit(as.character(KEGGoutput$Gene.IDs), "/"))
KEGGoutput_unnest                     <- unnest(KEGGoutput, Gene.IDs)
KEGGoutput_unnest$Cvirginica_KEGGID   <- paste("cvn:", KEGGoutput_unnest$Gene.IDs, sep='')
KEGGoutput_allgenes                   <- merge(KEGGoutput_unnest, Airr_Cvirg_Cgig_KEGG, by='Cvirginica_KEGGID') %>% 
                                                                       group_by(pathway.name) %>% 
                                                                       arrange(ProteinID, .by_group = TRUE) %>%
                                                                       dplyr::select(!c(TranscriptID,
                                                                                        Cvirginica_TranscriptID,
                                                                                        Airradians_TranscriptID)) %>% 
                                                                       unique()

write.csv(KEGGoutput_allgenes, file = paste("DESeq2/Venn_intersections/clusterProfiler/Low_expression_ModeratepCO2_KEGG_unlisted.csv", sep ='')) 

```

### Low Expression by HighpCO2h (low v. high and moderate v. high, overlapped DEGs for a sureg pattern by high in both models)

**input:** CohortRes_LowExp.HighpCO2h

```{r Cohort Low Expression by HighpCO2h  DESEq2 results - run KEGG}

entrezID_vector <- as.vector(as.numeric(gsub('.*cvn:', '',CohortRes_LowExp.HighpCO2h$Cvirginica_KEGGID)))
NCBIGeneID_vector <- as.vector(as.numeric(gsub('.*LOC', '',CohortRes_LowExp.HighpCO2h$blastxEval_CvirgProteinID)))
KEGG_cvirg     <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              #keyType="ncbi-geneid",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05) 
df              <- as.data.frame(head(KEGG_cvirg))
rownames(df)    <- c()
KEGGoutput      <- as.data.frame(do.call(cbind.data.frame, df)) %>% 
                                                dplyr::mutate(
                                                              Rich_factor  = (  (as.numeric(sub("/.*", "", GeneRatio))) / 
                                                                          (as.numeric(sub("/.*", "", BgRatio)))),
                                                              Type = 'Cohort low expression moderate pCO2') %>% 
                                                dplyr::rename(KEGGID_pathway = ID,
                                                              pathway.name   = Description,
                                                              Gene.IDs       = geneID,
                                                              Num.genes.exp  = Count) %>% 
                                                dplyr::mutate(Num.genes.all  = (as.numeric(sub("/.*", "", BgRatio))), 
                                                              log10_pvalue   = abs(log10(pvalue)),
                                                              pathway.name = sapply(strsplit(pathway.name, " - Crassostrea"),
                                                                                    "[",1)) %>% # ommit the unneeded pathway string
                                                dplyr::select('Type', 'KEGGID_pathway', 'pathway.name',
                                                              'Num.genes.all', 'Num.genes.exp', 'Gene.IDs',
                                                              'Rich_factor', 'pvalue', 'log10_pvalue', 'qvalue') %>% 
                                                arrange(desc(as.numeric(Rich_factor))) 
                    
write.csv(KEGGoutput, file = paste("DESeq2/Venn_intersections/clusterProfiler/Low_expression_HighpCO2_KEGG.csv", sep ='')) 
                    
# Plot
plot<- KEGGoutput %>%  
          ggplot(aes(x=reorder(pathway.name, Rich_factor), y= Rich_factor)) + 
          geom_point( aes(col=qvalue, size=Num.genes.exp)) +   # Draw points
          geom_segment(aes(x=pathway.name, 
                           xend=pathway.name, 
                           y=min(Rich_factor), 
                           yend=max(Rich_factor)),  
                           linetype=NA, 
                           size=0) +   # Draw dashed lines
          labs(title="Low Expression High pCO2h", 
               x = "Pathway",
               y = "Rich Factor") +
          theme_bw() +
          coord_flip()
pdf(paste("DESeq2/Venn_intersections/clusterProfiler/Low_expression_HighpCO2_RichFactorPlot.pdf", sep =''),width=8, height=6)
print(plot)
dev.off()
                     
# stringsplit and unnest for a data set of genes and IDs associated with each pathway 
KEGGoutput$Gene.IDs  <- as.vector(strsplit(as.character(KEGGoutput$Gene.IDs), "/"))
KEGGoutput_unnest                     <- unnest(KEGGoutput, Gene.IDs)
KEGGoutput_unnest$Cvirginica_KEGGID   <- paste("cvn:", KEGGoutput_unnest$Gene.IDs, sep='')
KEGGoutput_allgenes                   <- merge(KEGGoutput_unnest, Airr_Cvirg_Cgig_KEGG, by='Cvirginica_KEGGID') %>% 
                                                                       group_by(pathway.name) %>% 
                                                                       arrange(ProteinID, .by_group = TRUE) %>%
                                                                       dplyr::select(!c(TranscriptID,
                                                                                        Cvirginica_TranscriptID,
                                                                                        Airradians_TranscriptID)) %>% 
                                                                       unique()

write.csv(KEGGoutput_allgenes, file = paste("DESeq2/Venn_intersections/clusterProfiler/Low_expression_HighpCO2_KEGG_unlisted.csv", sep ='')) 

```

## (2) Challenge (WGCNA, frontloaded and costly activation)

### Frontloaded genes - combined modules pink and magneta from Challenge

**input:** ChallengeRes_Frontloaded - genes following the consitutige frontloaded criteria from modules pink + magenta 

```{r Frontloaded genes  WGCNA results - run KEGG}

entrezID_vector   <- as.vector(as.numeric(gsub('.*cvn:', '',ChallengeRes_Frontloaded$Cvirginica_KEGGID)))
NCBIGeneID_vector <- as.vector(as.numeric(gsub('.*LOC', '',ChallengeRes_Frontloaded$Cvirginica_GeneID)))
KEGG_cvirg     <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05) 
df              <- as.data.frame(head(KEGG_cvirg))
rownames(df)    <- c()
KEGGoutput      <- as.data.frame(do.call(cbind.data.frame, df)) %>% 
                                                dplyr::mutate(
                                                              Rich_factor  = (  (as.numeric(sub("/.*", "", GeneRatio))) / 
                                                                          (as.numeric(sub("/.*", "", BgRatio)))),
                                                              Type = 'Frontloaded',
                                                              modColor = 'pink and magenta') %>% 
                                                dplyr::rename(KEGGID_pathway = ID,
                                                              pathway.name   = Description,
                                                              Gene.IDs       = geneID,
                                                              Num.genes.exp  = Count) %>% 
                                                dplyr::mutate(Num.genes.all  = (as.numeric(sub("/.*", "", BgRatio))), 
                                                              log10_pvalue   = abs(log10(pvalue)),
                                                              pathway.name = sapply(strsplit(pathway.name, " - Crassostrea"),
                                                                                    "[",1)) %>% # ommit the unneeded pathway string
                                                dplyr::select('Type', 'modColor', 'KEGGID_pathway', 'pathway.name',
                                                              'Num.genes.all', 'Num.genes.exp', 'Gene.IDs',
                                                              'Rich_factor', 'pvalue', 'log10_pvalue', 'qvalue') %>% 
                                                arrange(desc(as.numeric(Rich_factor))) 
                    
write.csv(KEGGoutput, file = paste("WGCNA/challenge/Frontloaded_Activated/clusterProfiler/PinkMagenta_Frontloaded_KEGG.csv", sep ='')) 
                    
# Plot
plot<- KEGGoutput %>%  
          ggplot(aes(x=reorder(pathway.name, Rich_factor), y= Rich_factor)) + 
          geom_point( aes(col=qvalue, size=Num.genes.exp)) +   # Draw points
          geom_segment(aes(x=pathway.name, 
                           xend=pathway.name, 
                           y=min(Rich_factor), 
                           yend=max(Rich_factor)),  
                           linetype=NA, 
                           size=0) +   # Draw dashed lines
          labs(title="Frontloaded genes", 
               x = "Pathway",
               y = "Rich Factor") +
          theme_bw() +
          coord_flip()
pdf(paste("WGCNA/challenge/Frontloaded_Activated/clusterProfiler/PinkMagenta_Frontloaded_RichFactorPlot.pdf", sep =''),width=8, height=6)
print(plot)
dev.off()
                     
# stringsplit and unnest for a data set of genes and IDs associated with each pathway 
KEGGoutput$Gene.IDs  <- as.vector(strsplit(as.character(KEGGoutput$Gene.IDs), "/"))
KEGGoutput_unnest                     <- unnest(KEGGoutput, Gene.IDs)
KEGGoutput_unnest$Cvirginica_KEGGID   <- paste("cvn:", KEGGoutput_unnest$Gene.IDs, sep='')
KEGGoutput_allgenes                   <- merge(KEGGoutput_unnest, 
                                                (ChallengeRes_Frontloaded %>% 
                                                   select(Protein_name, Cvirginica_KEGGID, Cvirginica_TranscriptID, Airradians_TranscriptID)),
                                                by='Cvirginica_KEGGID') %>% 
                                                                       group_by(pathway.name) %>% 
                                                                       arrange(Protein_name, .by_group = TRUE) %>%
                                                                       unique()

write.csv(KEGGoutput_allgenes, file = paste("WGCNA/challenge/Frontloaded_Activated/clusterProfiler/PinkMagenta_Frontloaded_KEGG_unlisted.csv", sep ='')) 

```


### Activated genes - combined modules pink and magneta from Challenge

**input:** ChallengeRes_Activated - genes following the activated criteria from modules pink + magenta 

```{r Activated genes  WGCNA results - run KEGG}

entrezID_vector <- as.vector(as.numeric(gsub('.*cvn:', '',ChallengeRes_Activated$Cvirginica_KEGGID)))
KEGG_cvirg      <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "none",
                              pvalueCutoff = 0.05) 
df              <- as.data.frame(head(KEGG_cvirg))
rownames(df)    <- c()
KEGGoutput      <- as.data.frame(do.call(cbind.data.frame, df)) %>% 
                                                dplyr::mutate(
                                                              Rich_factor  = (  (as.numeric(sub("/.*", "", GeneRatio))) / 
                                                                          (as.numeric(sub("/.*", "", BgRatio)))),
                                                              Type = 'Frontloaded',
                                                              modColor = 'pink and magenta') %>% 
                                                dplyr::rename(KEGGID_pathway = ID,
                                                              pathway.name   = Description,
                                                              Gene.IDs       = geneID,
                                                              Num.genes.exp  = Count) %>% 
                                                dplyr::mutate(Num.genes.all  = (as.numeric(sub("/.*", "", BgRatio))), 
                                                              log10_pvalue   = abs(log10(pvalue)),
                                                              pathway.name = sapply(strsplit(pathway.name, " - Crassostrea"),
                                                                                    "[",1)) %>% # ommit the unneeded pathway string
                                                dplyr::select('Type', 'modColor', 'KEGGID_pathway', 'pathway.name',
                                                              'Num.genes.all', 'Num.genes.exp', 'Gene.IDs',
                                                              'Rich_factor', 'pvalue', 'log10_pvalue', 'qvalue') %>% 
                                                arrange(desc(as.numeric(Rich_factor))) 
                    
write.csv(KEGGoutput, file = paste("WGCNA/challenge/Frontloaded_Activated/clusterProfiler/PinkMagenta_Activated_KEGG.csv", sep ='')) 
                    
# Plot
plot<- KEGGoutput %>%  
          ggplot(aes(x=reorder(pathway.name, Rich_factor), y= Rich_factor)) + 
          geom_point( aes(col=qvalue, size=Num.genes.exp)) +   # Draw points
          geom_segment(aes(x=pathway.name, 
                           xend=pathway.name, 
                           y=min(Rich_factor), 
                           yend=max(Rich_factor)),  
                           linetype=NA, 
                           size=0) +   # Draw dashed lines
          labs(title="Frontloaded genes", 
               x = "Pathway",
               y = "Rich Factor") +
          theme_bw() +
          coord_flip()
pdf(paste("WGCNA/challenge/Frontloaded_Activated/clusterProfiler/PinkMagenta_Activated_RichFactorPlot.pdf", sep =''),width=8, height=6)
print(plot)
dev.off()
                     
# stringsplit and unnest for a data set of genes and IDs associated with each pathway 
KEGGoutput$Gene.IDs  <- as.vector(strsplit(as.character(KEGGoutput$Gene.IDs), "/"))
KEGGoutput_unnest                     <- unnest(KEGGoutput, Gene.IDs)
KEGGoutput_unnest$Cvirginica_KEGGID   <- paste("cvn:", KEGGoutput_unnest$Gene.IDs, sep='')
KEGGoutput_allgenes                   <- merge(KEGGoutput_unnest, 
                                                (ChallengeRes_Activated %>% 
                                                   select(Protein_name, Cvirginica_KEGGID, Cvirginica_TranscriptID, Airradians_TranscriptID)),
                                                by='Cvirginica_KEGGID') %>% 
                                                                       group_by(pathway.name) %>% 
                                                                       arrange(Protein_name, .by_group = TRUE) %>%
                                                                       unique()

write.csv(KEGGoutput_allgenes, file = paste("WGCNA/challenge/Frontloaded_Activated/clusterProfiler/PinkMagenta_Activat_KEGG_unlisted.csv", sep ='')) 

```


### Costly Activated genes - combined modules red and brown from Challenge

**input:** ChallengeRes_CostlyActivation - genes following the consitutige frontloaded criteria from modules pink + magenta 

```{r Costly Activation genes  WGCNA results - run KEGG}

entrezID_vector <- as.vector(as.numeric(gsub('.*cvn:', '',ChallengeRes_CostlyActivation$Cvirginica_KEGGID)))
KEGG_cvirg      <- enrichKEGG(gene = entrezID_vector, 
                              organism  = 'cvn', # 'hsa' is human 'crg' is pacific oyster 
                              keyType="kegg",
                              pAdjustMethod = "BH",
                              pvalueCutoff = 0.05) 
#...0 enriched terms found
```
